check that dsviews are all soft deleted

+++ front/lib/resources/data_source_resource.ts
@@ -32,7 +32,7 @@ import type {
   Result,
   UserType,
 } from "@app/types";
-import { formatUserFullName, Ok, removeNulls } from "@app/types";
+import { Err, formatUserFullName, Ok, removeNulls } from "@app/types";
 
 import { DataSourceViewModel } from "./storage/models/data_source_view";
 
@@ -399,6 +399,19 @@ export class DataSourceResource extends ResourceWithSpace<DataSourceModel> {
     transaction?: Transaction
   ): Promise<Result<number, Error>> {
     // We assume the data source views are already soft-deleted here.
+    const dataSourceViews = await DataSourceViewModel.findAll({
+      where: {
+        workspaceId: auth.getNonNullableWorkspace().id,
+        dataSourceId: this.id,
+        deletedAt: {
+          [Op.is]: null,
+        },
+      },
+    });
+
+    if (dataSourceViews.length > 0) {
+      return new Err(new Error("Data source views still exist"));
+    }
 
     const deletedCount = await this.model.destroy({
       where: {
